- Refactor the execution of RebaseCurrentToParent into a Task.Backgroundable - using it in TraverseAction
