+++ frontend/actions/src/main/java/com/virtuslab/gitmachete/frontend/actions/base/BaseSyncToParentByRebaseAction.java
@@ -8,39 +8,19 @@ import static io.vavr.API.Case;
 import static io.vavr.API.Match;
 
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.concurrent.atomic.AtomicReference;
 
 import com.intellij.dvcs.repo.Repository;
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.progress.ProgressIndicator;
-import com.intellij.openapi.progress.Task;
-import com.intellij.openapi.project.Project;
-import com.intellij.openapi.vcs.VcsNotifier;
-import git4idea.branch.GitRebaseParams;
-import git4idea.config.GitVersion;
-import git4idea.rebase.GitRebaseOption;
-import git4idea.rebase.GitRebaseUtils;
-import git4idea.repo.GitRepository;
-import git4idea.util.GitFreezingProcess;
-import io.vavr.control.Option;
-import io.vavr.control.Try;
 import kr.pe.kwonnam.slf4jlambda.LambdaLogger;
 import lombok.CustomLog;
 import lombok.experimental.ExtensionMethod;
 import lombok.val;
 import org.checkerframework.checker.guieffect.qual.UIEffect;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
-import com.virtuslab.gitmachete.backend.api.IGitMacheteRepositorySnapshot;
-import com.virtuslab.gitmachete.backend.api.IGitRebaseParameters;
 import com.virtuslab.gitmachete.backend.api.INonRootManagedBranchSnapshot;
-import com.virtuslab.gitmachete.backend.api.hooks.IExecutionResult;
-import com.virtuslab.gitmachete.frontend.actions.contextmenu.CheckoutSelectedAction;
 import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeyGitMacheteRepository;
 import com.virtuslab.gitmachete.frontend.defs.ActionPlaces;
 import com.virtuslab.gitmachete.frontend.resourcebundles.GitMacheteBundle;
-import com.virtuslab.qual.guieffect.UIThreadUnsafe;
 
 @ExtensionMethod({Arrays.class, GitMacheteBundle.class})
 @CustomLog
@@ -159,130 +139,12 @@ public abstract class BaseSyncToParentByRebaseAction extends BaseGitMacheteRepos
         && state != null && Repository.State.DETACHED == state;
 
     if (gitRepository != null && gitMacheteRepositorySnapshot != null) {
-      doRebase(project, gitRepository, gitMacheteRepositorySnapshot, branchToRebase, shouldExplicitlyCheckout);
+      new RebaseOnParentBackgroundable(project,
+          getString("action.GitMachete.BaseSyncToParentByRebaseAction.hook.task-title"),
+          gitRepository, gitMacheteRepositorySnapshot,
+          branchToRebase,
+          shouldExplicitlyCheckout).queue();
     }
   }
 
-  private void doRebase(
-      Project project,
-      GitRepository gitRepository,
-      IGitMacheteRepositorySnapshot gitMacheteRepositorySnapshot,
-      INonRootManagedBranchSnapshot branchToRebase,
-      boolean shouldExplicitlyCheckout) {
-    LOG.debug(() -> "Entering: project = ${project}, gitRepository = ${gitRepository}, branchToRebase = ${branchToRebase}");
-
-    val tryGitRebaseParameters = Try.of(branchToRebase::getParametersForRebaseOntoParent);
-
-    if (tryGitRebaseParameters.isFailure()) {
-      val e = tryGitRebaseParameters.getCause();
-      val message = e.getMessage() == null ? "Unable to get rebase parameters." : e.getMessage();
-      LOG.error(message);
-      VcsNotifier.getInstance(project).notifyError(/* displayId */ null,
-          getString("action.GitMachete.BaseSyncToParentByRebaseAction.notification.title.rebase-fail"), message);
-      return;
-    }
-
-    val gitRebaseParameters = tryGitRebaseParameters.get();
-    LOG.debug(() -> "Queuing machete-pre-rebase hooks background task for '${branchToRebase.getName()}' branch");
-
-    new Task.Backgroundable(project, getString("action.GitMachete.BaseSyncToParentByRebaseAction.hook.task-title")) {
-      @Override
-      @UIThreadUnsafe
-      public void run(ProgressIndicator indicator) {
-
-        final AtomicReference<Try<@Nullable IExecutionResult>> wrapper = new AtomicReference<>(Try.success(null));
-        new GitFreezingProcess(project, getTitle(), () -> {
-          LOG.info("Executing machete-pre-rebase hooks");
-          Try<@Nullable IExecutionResult> hookResult = Try
-              .of(() -> gitMacheteRepositorySnapshot.executeMachetePreRebaseHookIfPresent(gitRebaseParameters));
-          wrapper.set(hookResult);
-        }).execute();
-        Try<@Nullable IExecutionResult> hookResult = wrapper.get();
-        if (hookResult == null) {
-          // Not really possible, it's here just to calm down Checker Framework.
-          return;
-        }
-
-        if (hookResult.isFailure()) {
-          val message = "machete-pre-rebase hooks refused to rebase ${NL}error: ${hookResult.getCause().getMessage()}";
-          LOG.error(message);
-          VcsNotifier.getInstance(project).notifyError(/* displayId */ null,
-              getString("action.GitMachete.BaseSyncToParentByRebaseAction.notification.title.rebase-abort"),
-              message);
-          return;
-        }
-
-        val maybeExecutionResult = hookResult.get();
-        if (maybeExecutionResult != null && maybeExecutionResult.getExitCode() != 0) {
-          val message = "machete-pre-rebase hooks refused to rebase (exit code ${maybeExecutionResult.getExitCode()})";
-          LOG.error(message);
-          val executionResult = maybeExecutionResult;
-          val stdoutOption = executionResult.getStdout();
-          val stderrOption = executionResult.getStderr();
-          VcsNotifier.getInstance(project).notifyError(/* displayId */ null,
-              getString("action.GitMachete.BaseSyncToParentByRebaseAction.notification.title.rebase-abort"), message
-                  + (!stdoutOption.isBlank() ? NL + "stdout:" + NL + stdoutOption : "")
-                  + (!stderrOption.isBlank() ? NL + "stderr:" + NL + stderrOption : ""));
-          return;
-        }
-
-        LOG.debug(() -> "Queuing rebase background task for '${branchToRebase.getName()}' branch");
-
-        new Task.Backgroundable(project, getString("action.GitMachete.BaseSyncToParentByRebaseAction.task-title")) {
-          @Override
-          @UIThreadUnsafe
-          public void run(ProgressIndicator indicator) {
-            val params = getIdeaRebaseParamsOf(gitRepository, gitRebaseParameters);
-            LOG.info("Rebasing '${gitRebaseParameters.getCurrentBranch().getName()}' branch " +
-                "until ${gitRebaseParameters.getForkPointCommit().getHash()} commit " +
-                "onto ${gitRebaseParameters.getNewBaseBranch().getName()}");
-
-            /*
-             * Git4Idea ({@link git4idea.rebase.GitRebaseUtils#rebase}) does not allow to rebase in detached head state.
-             * However, it is possible with Git (performing checkout implicitly) and should be allowed in the case of
-             * "Checkout and Rebase Onto Parent" Action. To pass the git4idea check in such a case we checkout the branch
-             * explicitly and then perform the actual rebase.
-             */
-            if (shouldExplicitlyCheckout) {
-              CheckoutSelectedAction.doCheckout(
-                  project, indicator, gitRebaseParameters.getCurrentBranch().getName(), gitRepository);
-            }
-            GitRebaseUtils.rebase(project, Collections.singletonList(gitRepository), params, indicator);
-          }
-        }.queue();
-      }
-
-      // TODO (#95): on success, refresh only sync statuses (not the whole repository). Keep in mind potential
-      // changes to commits (eg. commits may get squashed so the graph structure changes).
-    }.queue();
-  }
-
-  @UIThreadUnsafe
-  private GitRebaseParams getIdeaRebaseParamsOf(GitRepository repository, IGitRebaseParameters gitRebaseParameters) {
-    val gitVersion = repository.getVcs().getVersion();
-    val currentBranchName = gitRebaseParameters.getCurrentBranch().getName();
-    val newBaseBranchFullName = gitRebaseParameters.getNewBaseBranch().getFullName();
-    val forkPointCommitHash = gitRebaseParameters.getForkPointCommit().getHash();
-
-    // TODO (#1114): remove the mechanism for checking the availability of "--empty=drop"
-    val options = kotlin.collections.SetsKt.hashSetOf(GitRebaseOption.INTERACTIVE);
-
-    val gitRebaseOption = getAvailableGitRebaseOptions("--empty=drop", gitVersion);
-    if (gitRebaseOption != null) {
-      options.add(gitRebaseOption);
-    }
-
-    return new GitRebaseParams(gitVersion, currentBranchName, newBaseBranchFullName,
-        /* upstream */ forkPointCommitHash, /* selectedOptions */ options, GitRebaseParams.AutoSquashOption.DEFAULT,
-        /* editorHandler */ null);
-  }
-
-  @UIThreadUnsafe
-  private @Nullable GitRebaseOption getAvailableGitRebaseOptions(String optionText, GitVersion gitVersion) {
-    val maybeEmptyDropEntry = GitRebaseOption.values().stream()
-        .filter(entry -> entry.getOption(gitVersion).equals(optionText))
-        .findFirst();
-
-    return Option.ofOptional(maybeEmptyDropEntry).getOrNull();
-  }
 }
